+++ Makefile
@@ -6,7 +6,7 @@ else
 PROTOBUF_ZIP = protoc-$(PROTOBUF_VERSION)-linux-x86_64.zip
 endif
 
-APIS = EngineAPI EnvironmentAPI DevTools
+APIS = EngineAPI EnvironmentAPI DevTools UserIdentity UserActionModule SocialController SignedFetch RestrictedActions PortableExperiences Players Permissions ParcelIdentity EthereumController CommunicationsController SceneStateStorageController
 
 NODE = node
 COMPILER = $(NODE) --max-old-space-size=4096 node_modules/.bin/decentraland-compiler
@@ -199,14 +199,14 @@ install_protoc:
 	chmod +x ./node_modules/.bin/protobuf/bin/protoc
 
 compile_apis: 
-	@echo "Generating APIs .proto ..." 
+	@echo "Generating APIs .proto ..." \
+	rm -rf ./packages/shared/apis/gen
+	mkdir -p ./packages/shared/apis/gen
 	@for file in $(APIS); do \
 		echo "Processing" $${file}; \
-		rm -rf ./packages/shared/apis/$${file}/gen; \
-		mkdir ./packages/shared/apis/$${file}/gen; \
 		./node_modules/.bin/protobuf/bin/protoc \
 			--plugin=./node_modules/.bin/protoc-gen-ts_proto \
 			--ts_proto_opt=esModuleInterop=true,returnObservable=false,outputServices=generic-definitions \
-			--ts_proto_out="./packages/shared/apis/$${file}/gen" -I="./packages/shared/apis/$${file}" \
-			"./packages/shared/apis/$${file}/$${file}.proto"; \
+			--ts_proto_out="./packages/shared/apis/gen" -I="./packages/shared/apis/" \
+			"./packages/shared/apis/$${file}.proto"; \
 	done
