--- packages/shared/apis/DevTools/client.ts
@@ -1,6 +1,0 @@
-import { loadService } from '@dcl/rpc/dist/codegen'
-import { RpcClientPort } from '@dcl/rpc/dist/types'
-import { DevToolsServiceDefinition } from './gen/DevTools'
-
-export const createDevToolsServiceClient = <Context>(clientPort: RpcClientPort) =>
-  loadService<Context, DevToolsServiceDefinition>(clientPort, DevToolsServiceDefinition)
