+++ packages/shared/apis/Empty.proto
@@ -1,0 +1,3 @@
+syntax = "proto3";
+
+message Empty {}
