--- packages/shared/apis/EngineAPI/client.ts
@@ -1,6 +1,0 @@
-import { loadService } from '@dcl/rpc/dist/codegen'
-import { RpcClientPort } from '@dcl/rpc/dist/types'
-import { EngineAPIServiceDefinition } from './gen/EngineAPI'
-
-export const createEngineAPIServiceClient = <Context>(clientPort: RpcClientPort) =>
-  loadService<Context, EngineAPIServiceDefinition>(clientPort, EngineAPIServiceDefinition)
