+++ packages/shared/apis/Players.proto
@@ -1,0 +1,27 @@
+syntax = "proto3";
+
+import "UserIdentity.proto";
+import "Empty.proto";
+
+
+message Player {
+    string user_id = 1;
+}
+
+message PlayersGetUserDataResponse {
+    optional UserData data = 1;
+}
+
+message PlayerListResponse {
+    repeated Player players = 1;
+}
+
+message GetPlayerDataRequest {
+    string userId = 1;
+}
+
+service PlayersService {
+    rpc getPlayerData(GetPlayerDataRequest) returns (PlayersGetUserDataResponse) {}
+    rpc getPlayersInScene(Empty) returns (PlayerListResponse) {}
+    rpc getConnectedPlayers(Empty)  returns (PlayerListResponse) {}
+}
