+++ packages/shared/apis/SocialController.proto
@@ -1,0 +1,7 @@
+syntax = "proto3";
+
+message Empty {}
+
+service SocialControllerService {
+    rpc test(Empty) returns (Empty) {}
+}
