+++ packages/shared/apis/UserActionModule.proto
@@ -1,0 +1,11 @@
+syntax = "proto3";
+
+message RequestTeleportRequest {
+    string destination = 1;
+}
+
+message Empty {}
+
+service UserActioNoduleService {
+    rpc requestTeleport(RequestTeleportRequest) returns (Empty) {}
+}
