+++ packages/shared/apis/UserIdentity.proto
@@ -1,0 +1,43 @@
+syntax = "proto3";
+
+import "Empty.proto";
+
+message Snapshots { 
+    string face256 = 1;
+    string body = 2;
+}
+
+message AvatarForUserData {
+    string body_shape = 1;
+    string skin_color = 2;
+    string hair_color = 3;
+    string eye_color = 4;
+    repeated string wearables = 5;
+    Snapshots snapshots = 6;
+}
+
+message UserData {
+    string display_name = 1;
+    optional string public_key = 2;
+    bool has_connected_web3 = 3;
+    string user_id = 4;
+    int32 version = 5;
+    AvatarForUserData avatar = 6; 
+}
+
+message GetUserPublicKeyData {
+    string address = 1;
+}
+
+message GetUserDataResponse {
+    optional UserData data = 1;
+}
+
+message GetUserPublicKeyResponse {
+    optional GetUserPublicKeyData data = 1;
+}
+
+service UserIdentityService {
+    rpc getUserPublicKey(Empty) returns (GetUserPublicKeyResponse) {}
+    rpc getUserData(Empty) returns (GetUserDataResponse) {}
+}
